+++ Backend/routes/biditemcontroller.js
@@ -1,0 +1,10 @@
+// routes/bidItemRoutes.js
+const express = require('express');
+const router = express.Router();
+const bidItemController = require('../controllers/biditemcontroller');
+
+router.put('/items', bidItemController.updateItemStatus);
+router.get('/biditems', bidItemController.getBidItems);
+router.post('/lots/:lot_no/bids', bidItemController.createBid);
+
+module.exports = router;
