+++ Backend/routes/userroutes.js
@@ -1,0 +1,8 @@
+// routes/userRoutes.js
+const express = require('express');
+const router = express.Router();
+const userController = require('../controllers/usercontroller');
+
+router.get('/users', userController.getAllUsers);
+
+module.exports = router;
